Split types to separate files
